refactor: Move string-to-decimal API to DecimalUtil
